Renaming to lowercase and eliminated simulation.py and combined it with input.py
